Move the log2 computation to a standalone object
